requirements.txt
aniso8601==9.0.1
attrs==21.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Flask==2.0.2
flask-restx==0.5.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
pycodestyle==2.8.0
pyrsistent==0.18.0
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
gunicorn==20.1.0
flask-cors==5.0.0
flask-jwt-extended==4.6.0